+++ packages/phone-number-privacy/package.json
@@ -18,7 +18,8 @@
     "db:migrate": "NODE_ENV=development ts-node scripts/runMigrations.ts"
   },
   "dependencies": {
-    "@celo/contractkit": "0.3.2-dev",
+    "@celo/contractkit": "0.3.2",
+    "@celo/utils": "0.1.8",
     "bls12377js-blind": "https://github.com/celo-org/bls12377js-blind#3a5ac033fbe282d40002639e01db56313124e721",
     "firebase-admin": "^8.9.2",
     "firebase-functions": "^3.3.0",
