+++ packages/phone-number-privacy/src/common/identity.ts
@@ -1,0 +1,7 @@
+/*
+ * Confirms that user is who they say they are and throws error on failure to confirm.
+ */
+export function confirmUser() {
+  //TODO [amyslawson]
+  return
+}
