+++ packages/phone-number-privacy/src/web3/contracts.ts
@@ -1,9 +1,7 @@
-import { newKitFromWeb3 } from '@celo/contractkit'
-import Web3 from 'web3'
+import { newKit } from '@celo/contractkit'
 import config from '../config'
 
-const web3: Web3 = new Web3(new Web3.providers.HttpProvider(config.blockchain.provider))
-const contractKit = newKitFromWeb3(web3)
+const contractKit = newKit(config.blockchain.provider)
 
 export function getContractKit() {
   return contractKit
